requirements.txt
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
fastapi==0.111.0
pydantic==2.8.0
uvicorn==0.30.1
PyJWT==2.8.0
pytest==8.3.0